+++ internal/home/client.go
@@ -3,8 +3,10 @@ package home
 import (
 	"encoding"
 	"fmt"
+	"time"
 
 	"github.com/AdguardTeam/AdGuardHome/internal/filtering"
+	"github.com/AdguardTeam/AdGuardHome/internal/filtering/safesearch"
 	"github.com/AdguardTeam/dnsproxy/proxy"
 )
 
@@ -41,6 +43,23 @@ func (c *Client) closeUpstreams() (err error) {
 			return fmt.Errorf("closing upstreams of client %q: %w", c.Name, err)
 		}
 	}
+
+	return nil
+}
+
+// setSafeSearch initializes and sets the safe search filter for this client.
+func (c *Client) setSafeSearch(
+	conf filtering.SafeSearchConfig,
+	cacheSize uint,
+	cacheTTL time.Duration,
+) (err error) {
+	ss, err := safesearch.NewDefault(conf, fmt.Sprintf("client %q", c.Name), cacheSize, cacheTTL)
+	if err != nil {
+		// Don't wrap the error, because it's informative enough as is.
+		return err
+	}
+
+	c.SafeSearch = ss
 
 	return nil
 }